Rename markup-translator to slacken
