+++ lib/slacken/slack_url.rb
@@ -1,4 +1,4 @@
-module MarkupTranslator
+module Slacken
   module SlackUrl
     def self.link_tag(title, url)
       "<#{url}|#{title}>"
