+++ lib/slacken/version.rb
@@ -1,3 +1,3 @@
-module MarkupTranslator
+module Slacken
   VERSION = "0.1.0"
 end
