+++ sample/source.html
@@ -1,5 +1,5 @@
 
-<h1>MarkupTranslator</h1>
+<h1>Slacken</h1>
 
 <p>This gem translates a html source into <strong>a markup text for Slack</strong>.<br>
 <a href="http://qiita.com">Qiita</a> uses this gem to decorate notification messages to Slack <img class="emoji" title=":trollface:" alt=":trollface:" src="https://cdn.qiita.com/emoji/trollface.png" height="20" width="20" align="absmiddle">.</p>
