+++ slacken.gemspec
@@ -1,16 +1,16 @@
 # -*- encoding: utf-8 -*-
 
-require File.expand_path('../lib/markup-translator/version', __FILE__)
+require File.expand_path('../lib/slacken/version', __FILE__)
 
 Gem::Specification.new do |gem|
-  gem.name          = "markup-translator"
-  gem.version       = MarkupTranslator::VERSION
+  gem.name          = "slacken"
+  gem.version       = Slacken::VERSION
   gem.summary       = %q{Translate HTML sources to markup texts for slack}
   gem.description   = %q{Translate HTML sources to markup texts for slack}
   gem.license       = "MIT"
   gem.authors       = ["Tomoya Chiba"]
   gem.email         = "tomo.asleep@gmail.com"
-  gem.homepage      = "https://rubygems.org/gems/markup-translator"
+  gem.homepage      = "https://rubygems.org/gems/slacken"
 
   gem.files         = `git ls-files`.split($/)
   gem.executables   = gem.files.grep(%r{^bin/}).map{ |f| File.basename(f) }
