+++ spec/slacken/document_component/sanitize_special_tag_containers_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-class MarkupTranslator::DocumentComponent
+class Slacken::DocumentComponent
   describe SanitizeSpecialTagContainers, dsl: true do
     describe '#sanitized?' do
       subject { component.sanitized? }
